requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==4.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
